chore: improve test organization
