+++ goopstest/status_test.go
@@ -43,120 +43,35 @@ func ConfigureMaintenance() error {
 	return nil
 }
 
-func ConfigureMaintenanceOnStart() error {
-	env := goops.ReadEnv()
-	if env.HookName == "start" {
-		err := goops.SetUnitStatus(goops.StatusMaintenance, "Performing maintenance")
-		if err != nil {
-			return err
-		}
-	} else {
-		err := goops.SetUnitStatus(goops.StatusActive, "Charm is active")
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func ConfigureMaintenanceIfLeader() error {
-	isLeader, err := goops.IsLeader()
-	if err != nil {
-		return err
-	}
-
-	if isLeader {
-		err := goops.SetUnitStatus(goops.StatusMaintenance, "Performing maintenance on leader")
-		if err != nil {
-			return err
-		}
-	} else {
-		err := goops.SetUnitStatus(goops.StatusActive, "Charm is active on non-leader")
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func ConfigureMaintenanceIfNotLeader() error {
-	isLeader, err := goops.IsLeader()
-	if err != nil {
-		return err
-	}
-
-	if !isLeader {
-		err := goops.SetUnitStatus(goops.StatusMaintenance, "Performing maintenance on leader")
-		if err != nil {
-			return err
-		}
-	} else {
-		err := goops.SetUnitStatus(goops.StatusActive, "Charm is active on non-leader")
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 func TestCharmStatus(t *testing.T) {
 	tests := []struct {
 		name     string
 		handler  func() error
 		hookName string
-		leader   bool
 		want     string
 	}{
 		{
 			name:     "ActiveStatus",
 			handler:  ConfigureActive,
 			hookName: "start",
-			leader:   false,
 			want:     string(goops.StatusActive),
 		},
 		{
 			name:     "BlockedStatus",
 			handler:  ConfigureBlocked,
 			hookName: "start",
-			leader:   false,
 			want:     string(goops.StatusBlocked),
 		},
 		{
 			name:     "WaitingStatus",
 			handler:  ConfigureWaiting,
 			hookName: "start",
-			leader:   false,
 			want:     string(goops.StatusWaiting),
 		},
 		{
 			name:     "MaintenanceStatus",
 			handler:  ConfigureMaintenance,
 			hookName: "start",
-			leader:   false,
-			want:     string(goops.StatusMaintenance),
-		},
-		{
-			name:     "MaintenanceStatusOnStart",
-			handler:  ConfigureMaintenanceOnStart,
-			hookName: "start",
-			leader:   false,
-			want:     string(goops.StatusMaintenance),
-		},
-		{
-			name:     "MaintenanceStatusIfLeader",
-			handler:  ConfigureMaintenanceIfLeader,
-			hookName: "start",
-			leader:   true,
-			want:     string(goops.StatusMaintenance),
-		},
-		{
-			name:     "MaintenanceStatusIfNotLeader",
-			handler:  ConfigureMaintenanceIfNotLeader,
-			hookName: "start",
-			leader:   false,
 			want:     string(goops.StatusMaintenance),
 		},
 	}
@@ -168,9 +83,7 @@ func TestCharmStatus(t *testing.T) {
 				Charm: tc.handler,
 			}
 
-			stateIn := &goopstest.State{
-				Leader: tc.leader,
-			}
+			stateIn := &goopstest.State{}
 
 			stateOut, err := ctx.Run(tc.hookName, stateIn)
 			if err != nil {
